feat: prepare omegaconf for multiprocessing
